api_calls/src/camunda-platform-rest-api-client/pyproject.toml
[tool.poetry]
name = "camunda-platform-rest-api-client"
version = "7.15.0"
description = "A client library for accessing Camunda Platform REST API"

authors = []

readme = "README.md"
packages = [
    {include = "camunda_platform_rest_api_client"},
]
include = ["CHANGELOG.md", "camunda_platform_rest_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.6"
httpx = ">=0.15.4,<0.20.0"
attrs = ">=20.1.0,<22.0.0"
python-dateutil = "^2.8.0"

[tool.black]
line-length = 120
target_version = ['py36', 'py37', 'py38']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"